Managed to get toggling of series monitoring working.
Example command:

    gonarr -c config.json -i 1 --season 1 -m

A great big caveat:

On the assumption that the server implementation was at least
minimally sane, I tried sending a subset of the whole JSON, with just
the id and seasons fields. This was rejected, so, guided by the error
responses, I added progressively more fields until the server accepted
the payload.  However, this turned out to be a really bad idea; it
wipes out all the other metadata. In other words, the server just
replaces the entire server-side state with whatever you give it; it
doesn't attempt to update the server-side state by "merging in" the
fields you provide. That's a really bad server-side implementation,
IMHO, but we have to work with it.